+++ src/pyiem/nws/products/_outlook_util.py
@@ -71,7 +71,7 @@ def ensure_outside_conus(ls):
         # Get new point that may be too close for comfort
         pt = get_conus_point(pt)
         if pt.within(CONUS["poly"]) or pt.distance(CONUS["poly"]) < 0.001:
-            LOG.info("     idx: %s is still within, evasive action", idx)
+            LOG.warning("     idx: %s is still within, evasive action", idx)
             done = False
             for multi in [0.01, 0.1, 1.0]:
                 if done:
@@ -90,10 +90,10 @@ def ensure_outside_conus(ls):
                     pt2 = translate(pt, xoff=xoff, yoff=yoff)
                     if not pt2.within(CONUS["poly"]):
                         pt = pt2
-                        LOG.info("     idx: %s is now %s", idx, pt)
+                        LOG.warning("     idx: %s is now %s", idx, pt)
                         done = True
                         break
-        LOG.info(
+        LOG.warning(
             "     fix idx: %s to new: %.4f %.4f Inside: %s",
             idx,
             pt.x,
@@ -116,7 +116,7 @@ def debug_draw(i, segment):
     ax.plot(segment[:, 0], segment[:, 1], c="b")
     ax.plot(CONUS["poly"].exterior.xy[0], CONUS["poly"].exterior.xy[1], c="r")
     mydir = tempfile.gettempdir()
-    LOG.info("writting %s/%sdebugdraw.png", mydir, i)
+    LOG.warning("writting %s/%sdebugdraw.png", mydir, i)
     fig.savefig(f"{mydir}/{i}debugdraw.png")
     return fig
 
@@ -126,7 +126,7 @@ def condition_segment(segment):
     # 1. If the start and end points are the same, done and one
     if segment[0][0] == segment[-1][0] and segment[0][1] == segment[-1][1]:
         if len(segment) == 2:
-            LOG.info("    REJECTING two point segment, both equal")
+            LOG.warning("    REJECTING two point segment, both equal")
             return None
         return [segment]
     # 2. If point start and end points are inside the conus and they are closer
@@ -139,7 +139,7 @@ def condition_segment(segment):
         cdist0 = cpt0.distance(pt0)
         cdist1 = cpt1.distance(pt1)
         if pt0.distance(pt1) < 0.5 * min([cdist0, cdist1]):
-            LOG.info("     non-closed polygon assumed unclosed in error.")
+            LOG.warning("     non-closed polygon assumed unclosed in error.")
             segment.append(segment[0])
             return [segment]
     # 3. If the line intersects the CONUS 3+ times, split the line
@@ -151,9 +151,9 @@ def condition_segment(segment):
     # We got multiple linestrings
     res = [r for r in res.geoms if r.length > 0.2]
     if len(res) == 1:
-        LOG.info("    was able to filter out very short lines")
+        LOG.warning("    was able to filter out very short lines")
         return [ensure_outside_conus(res[0]).coords]
-    LOG.info("     returning a MultiLineString len=%s", len(res))
+    LOG.warning("     returning a MultiLineString len=%s", len(res))
     return [ensure_outside_conus(x).coords for x in res]
 
 
@@ -224,11 +224,11 @@ def winding_logic(linestrings):
                 ]
             LOG.debug("     found %s filtered rows", len(df2.index))
             if df2.empty:
-                LOG.info("     i=%s adding poly: %.3f", i, poly.area)
+                LOG.warning("     i=%s adding poly: %.3f", i, poly.area)
                 if poly not in polys:
                     polys.append(poly)
                 else:
-                    LOG.info("     this polygon is a dup, skipping")
+                    LOG.warning("     this polygon is a dup, skipping")
                 break
             # updated ended_at
             ended_at = df2.iloc[0]["end"]
@@ -252,7 +252,9 @@ def sql_day_collect(prod, txn, day, collect):
             "DELETE from spc_outlook_geometries where spc_outlook_id = %s",
             (outlook_id,),
         )
-        LOG.info("Removed %s rows from spc_outlook_geometries", txn.rowcount)
+        LOG.warning(
+            "Removed %s rows from spc_outlook_geometries", txn.rowcount
+        )
         # Update the updated column
         txn.execute(
             "UPDATE spc_outlook SET updated = now() WHERE id = %s",
@@ -308,13 +310,13 @@ def _sql_cycle_canonical(prod, txn, day, collect, outlook_id):
         (collect.expire, prod.outlook_type, day, prod.cycle),
     )
     if txn.rowcount == 0:  # yes
-        LOG.info("Setting as canonical cycle of %s", prod.cycle)
+        LOG.warning("Setting as canonical cycle of %s", prod.cycle)
         _sql_set_cycle(txn, outlook_id, prod.cycle)
     else:
         # tricky
         for row in txn.fetchall():
             if row["product_issue"] < prod.valid:
-                LOG.info(
+                LOG.warning(
                     "Setting old outlook %s to cycle=-1, product_issue = %s "
                     ", prod.valid = %s",
                     row["id"],
@@ -324,14 +326,14 @@ def _sql_cycle_canonical(prod, txn, day, collect, outlook_id):
                 _sql_set_cycle(txn, row["id"], -1)
             elif row["product_issue"] > prod.valid:
                 prod.cycle = -1
-        LOG.info("Setting this outlook to cycle=%s", prod.cycle)
+        LOG.warning("Setting this outlook to cycle=%s", prod.cycle)
         _sql_set_cycle(txn, outlook_id, prod.cycle)
 
 
 def quality_control(prod):
     """Do Quality Control work."""
     # 1. Do polygons overlap for the same outlook
-    LOG.info("==== Running Quality Control Checks")
+    LOG.warning("==== Running Quality Control Checks")
     for day, collect in prod.outlook_collections.items():
         # Everything should be smaller than General Thunder, for conv
         tstm = prod.get_outlook("CATEGORICAL", "TSTM", day)
@@ -345,7 +347,7 @@ def quality_control(prod):
                         f"Area: {outlook.geometry.area:.2f} "
                         f"TSTM Area: {tstm.geometry.area:.2f}"
                     )
-                    LOG.info(msg)
+                    LOG.warning(msg)
                     prod.warnings.append(msg)
                     continue
                 if poly.area < 0.1:
@@ -353,7 +355,7 @@ def quality_control(prod):
                         f"Impossibly small polygon.area {poly.area:.2f} "
                         "discarded"
                     )
-                    LOG.info(msg)
+                    LOG.warning(msg)
                     prod.warnings.append(msg)
                     continue
                 intersect = CONUS["poly"].intersection(poly)
@@ -377,14 +379,14 @@ def quality_control(prod):
                     if intersection.is_empty:
                         continue
                     if intersection.area < 0.1:
-                        LOG.info("Ignoring small intersection of polygons")
+                        LOG.warning("Ignoring small intersection of polygons")
                         continue
                     passes_check = False
                     msg = (
                         f"Discarding polygon idx: {i} as it intersects "
                         f"idx: {i2} Area: {poly.area:.2f}"
                     )
-                    LOG.info(msg)
+                    LOG.warning(msg)
                     prod.warnings.append(msg)
                     break
                 if passes_check: